Bug 1696: Updated test-programs and fixed very old bug in PVSSinfo::typeExists. RTDBPort works ok except that the writer will receive it own event is has send. Have to introduce seqnrs.

+++ branches/TBBExtTrigger-Bug1696-task-branch/MAC/GCF/PVSS/test/tPVSSservice.cc
@@ -78,7 +78,7 @@ GCFEvent::TResult tGSAService::initial(GCFEvent& e, GCFPortInterface& /*p*/)
 		itsService = new PVSSservice(itsResponse);
 
 		// test PVSSInfo class
-		int8	sysID;
+		int	sysID;
 		string	sysName;
 		LOG_DEBUG_STR("LocalSystemName: " << PVSSinfo::getLocalSystemName());
 		LOG_DEBUG_STR("LocalSystemID  : " << (sysID = PVSSinfo::getLocalSystemId()));
